backend/pyproject.toml
[project]
name = "hyper-scale-core-async-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-prometheus>=0.60b1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "prometheus-client>=0.24.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
